timestamp-service: don't restore time on shutdown
load-timestamp.sh --save should only save the current systemtime and not
evaluate /etc/timestamp and change the systemtime.
If the systemtime has been changed to something in the past of /etc/timestamp
then that change should be accepted without further checks.

While at it correct the comment. The RTC/hwclock is not touched by the
timestamp service. This is a copy paste error from the init.d implementation
in openembedded-core/meta/recipes-core/initscripts/initscripts-1.0/bootmisc.sh

Signed-off-by: Max Krummenacher <max.oss.09@gmail.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>