angstrom-core-tweaks.inc: actually use fs-perms-angstrom.txt
FILESYSTEM_PERMS_TABLES must be set with the relative path from one of the
BBPATH paths.

Without this patch the setting of FILESYSTEM_PERMS_TABLES disables the use
of openembedded-core/meta/files/fs-perms.txt but, as the file in
meta-angstrom/files/fs-perms-angstrom.txt is not found, that file is not
used for the fixup-perms task.

With fs-perms-angstrom.txt active files deployed from copydebugsources no
longer trigger these warnings:
| WARNING: helloworld-1.0-r0 do_package_qa: QA Issue: helloworld: .../src/main.c
| is owned by uid 1000, which is the same as the user running bitbake. This may
| be due to host contamination [host-user-contaminated]

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>