angstrom.inc: set arm instruction set with existing machine overrides
armv4t, armv5te are not (no longer?) set as MACHINEOVERRIDES if an armv4t
or armv5te tune is set.
Thus the default ARM_INSTRUCTION_SET ??= "thumb" kicks in.

Set the override to armv4, armv5 to get the expected ARM_INSTRUCTION_SET.

While at it cleanup comment to keep up with changes.

Build tested with a build for a armv5te machine.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>